+++ test/unit/rack/roger_test.rb
@@ -19,6 +19,25 @@ module Roger
         response = request.get("/erb")
 
         assert response.body.include?("ERB format")
+        assert_equal "Roger", response.headers["X-Handled-By"]
+      end
+
+      def test_middleware_does_not_render_unrenderables
+        @project.construct.file "html/myjpeg.jpg", "JPG"
+        request = ::Rack::MockRequest.new(@app)
+        response = request.get("/myjpeg.jpg")
+
+        assert response.body.include?("JPG")
+        assert_equal nil, response.headers["X-Handled-By"]
+      end
+
+      def test_middleware_does_not_render_unwanteds
+        @project.construct.file "html/mysass.scss", ".scss{}"
+        request = ::Rack::MockRequest.new(@app)
+        response = request.get("/mysass.scss")
+
+        assert response.body.include?(".scss{}")
+        assert_equal nil, response.headers["X-Handled-By"]
       end
 
       def test_renderer_options_are_passed